+++ src/move.c
@@ -6,7 +6,7 @@
 /*   By: mcygan <mcygan@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/12 12:03:55 by mcygan            #+#    #+#             */
-/*   Updated: 2025/02/12 19:00:01 by mcygan           ###   ########.fr       */
+/*   Updated: 2025/02/14 16:00:33 by mcygan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 static void	move_w(t_data *data)
 {
-	float	new_x;
-	float	new_y;
+	double	new_x;
+	double	new_y;
 
 	new_x = data->player_x + cos(data->player_a) * MOVE_SPEED;
 	new_y = data->player_y + sin(data->player_a) * MOVE_SPEED;
@@ -27,8 +27,8 @@ static void	move_w(t_data *data)
 
 static void	move_a(t_data *data)
 {
-	float	new_x;
-	float	new_y;
+	double	new_x;
+	double	new_y;
 
 	new_x = data->player_x + cos(data->player_a - M_PI / 2) * LATERAL_SPEED;
 	new_y = data->player_y + sin(data->player_a - M_PI / 2) * LATERAL_SPEED;
@@ -40,8 +40,8 @@ static void	move_a(t_data *data)
 
 static void	move_s(t_data *data)
 {
-	float	new_x;
-	float	new_y;
+	double	new_x;
+	double	new_y;
 
 	new_x = data->player_x - cos(data->player_a) * MOVE_SPEED;
 	new_y = data->player_y - sin(data->player_a) * MOVE_SPEED;
@@ -53,8 +53,8 @@ static void	move_s(t_data *data)
 
 static void	move_d(t_data *data)
 {
-	float	new_x;
-	float	new_y;
+	double	new_x;
+	double	new_y;
 
 	new_x = data->player_x + cos(data->player_a + M_PI / 2) * LATERAL_SPEED;
 	new_y = data->player_y + sin(data->player_a + M_PI / 2) * LATERAL_SPEED;
